The column types should be chosen from the set of types that can actually be returned by the server (i.e. mediumint instead of integer).

+++ sql/src/odbc/driver/SQLGetTypeInfo.c
@@ -51,7 +51,7 @@ static const char *columnnames[NCOLUMNS] = {
 static const char *columntypes[NCOLUMNS] = {
 	"varchar",
 	"smallint",
-	"integer",
+	"mediumint",
 	"varchar",
 	"varchar",
 	"varchar",
@@ -66,7 +66,7 @@ static const char *columntypes[NCOLUMNS] = {
 	"smallint",
 	"smallint",
 	"smallint",
-	"integer",
+	"mediumint",
 	"smallint",
 };
